+++ modules/canvaskit/npm_build/types/canvaskit-wasm-tests.ts
@@ -560,6 +560,11 @@ function paragraphTests(CK: CanvasKit, p?: Paragraph) {
     const m = p.getLineMetrics(); // $ExpectType LineMetrics[]
     const n = CK.GlyphRunFlags.IsWhiteSpace === 1;
     const o = p.unresolvedCodepoints(); // $ExpectType number[]
+    const q = p.getLineMetricsAt(0); // $ExpectType LineMetrics | null
+    const r = p.getNumberOfLines(); // $ExpectType number
+    const s = p.getLineNumberAt(0); // $ExpectType number
+    const t = p.getGlyphInfoAt(0);  //$ExpectType GlyphInfo | null
+    const u = p.getClosestGlyphInfoAtCoordinate(10, 3);  //$ExpectType GlyphInfo | null
 }
 
 function paragraphBuilderTests(CK: CanvasKit, fontMgr?: FontMgr, paint?: Paint) {